Move healthz into util
